+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Latest
 -------------------
 
 * Officially support Python 3.12.
+* Fix: Error when using `click` version 6.0 and 7.0 (#191).
 
 1.11.1 (2023-08-21)
 -------------------